refactor: refatoracao dos components. v1
