+++ src/hooks/index.js
@@ -1,0 +1,3 @@
+import useMedia from './useMedia';
+
+export { useMedia };
