+++ src/pages/Home/containers/OurActions/Mobile/SlideMobile.style.js
