+++ src/pages/Home/containers/index.js
@@ -2,7 +2,7 @@ import AboutUs from './AboutUs/AboutUs';
 import Partners from './Partners/Partners';
 import WhereAreWe from './WhereAreWe/WhereAreWe';
 import SlideMobile from './OurActions/Mobile/SlideMobile';
-import SliderDesktop from './OurActions/Desktop/SliderDesktop';
+import SlideDesktop from './OurActions/Desktop/SlideDesktop';
 import Instagram from './Instagram/Instagram';
 import Banner from './Banner/Banner';
 
@@ -10,7 +10,7 @@ export {
   AboutUs,
   Partners,
   WhereAreWe,
-  SliderDesktop,
+  SlideDesktop,
   SlideMobile,
   Instagram,
   Banner,
